Refactors the core
All plugins write into `options` so their defaults might be get lost on
a refresh. For that `options` and `_options` are renamed to `settings`
and `options` to make the pattern a little bit more clear (jQuery does
it also in the docs examples). All plugins write their defaults into
`options` before options will merged into `settings` with responsive
options.

The position values (`this.pos.*`) replaced by a cleaner set of
functions to make the code hopefully more readable and devide the logic
into smaller peaces where it belongs to.

The `setSpeed` has renamed to `speed` and its calculation is in a
separate function `duration`. `speed` itself is now a simple
getter/setter.

The `animStage` has renamed to `animate` and uses `speed` togehter with
transitions. Also jQuery 1.8 supports vendor prefixes, thus this drop
suppert for anything below jQuery 1.8.3.

Starts using `_` as prefix for protected members which should not used
directly whenver it is not neccessary, espacially in plugins.

Corrects `trigger` what retriggers triggerable event. This also
introduces a small API to suppress and releases events, instead of
boolean parameters in different functions like `bypassEvent`.