Corrects `normalize`, `maximum` and `to`
The implementation of `normalize` is mixed up with that of `to`, which
needs to be seperated.

The maximum is wrongly implemented and doesn't respect the severeal
cases.

The option `navRewind` belongs to the core because it changes the way
how `to` behaves and it's often needed even when the `navigation` plugin
is not.