Corrects `destroy`
If a carousel is destroyed, any markup that was created during runtime
must be removed. Because the classes `owl-carousel` and `owl-theme` are
both added by the plugin, but are also commonly set by the user, here is
a distinction no longer possible. Therefore, they must always be set
manually by the user in the future. As part of this, all other classes
have been made ​​configurable. Also the class `owl-refresh` has been
moved into the plugin element and the `owl-drag` class replaces the
`owl-text-select-on` class for proper semantics.

Closes #298