Changes support plugins API
It seems to be pretty common to extend jQuery's `support` utility (e.g.
Twitter Bootstrap and jQuery Transit). Regarding to the documentation
(http://api.jquery.com/jQuery.cssHooks/) `$.support.<css>` should be set
to a string of the vendor-prefixed version of `<css>`. For `transtition`
Twitter Bootstrap doesn't respect this convention but this was easily
fixed with a string object.

The default support plugin has been leaved as simple like before to keep
the maintenance easy. For a better feature detection the Modernizr
support plugin should be used.

Refers #352