Removes calculated stage width
It seems that the explicit calculation and placing of the stage width is
not necessary at all. Similar can be easily achieved by proper CSS
styles. For the stage itself this is `white-space: nowrap` and for the
items it's `display: inline-block`.

Tested on Chrome, Firefox and IE. Hopefully I didn't forget anything.

Refers #346