Adds status plugin
Moves the status information of events into a separate plugin. The
property names for these status information has changed. From
`item.count` to `itemCount` and from `item.index` to `currentPosition`.
This is because the term `item` will be used for events like `prepare`,
`add` and `remove`. Also the `currentPosition` is now relative.

Refers #401