Fix infinite loop
This is a bit of an edge case, but it is possible to set the [`items` options](https://owlcarousel2.github.io/OwlCarousel2/docs/api-options.html) to a fractional number (e.g. 2.5 items per page instead of 3). This actually works well, except for these lines here. For example, if `items` is set to 2.5, and there are 4 images in the carousel, then `repeat` becomes 1.5. The first `repeat--` makes `repeat` equal .5, then the next loop `repeat` equals -.5. Because the loop is supposed to terminate when `repeat` is 0 (falsy), the loop instead continues to run, and `repeat` goes down to negative infinity. Eventually the browser runs out of memory with strange errors (Chrome just says "Invalid string length"). The proposed change fixes this problem.