Enhancement: add support for custom selectors for dots navigation
Implements #2077.
The problem: custom navigation set up via `dotsContainer` would fail to work
if the elements nesting is more than 2 in the "dots" elements, let alone
using <button> elements was required.
This commit introduces the new `dotSelector` option to the plugin, as well as
the new logic for what elements the plugin should process clicks for.
- If both `dotsContainer` and `dotSelector` are set, then the plugin will
listen for clicks on elements that match the selector from `dotSelector` and
are inside of `dotsContainer`.
- If `dotsContainer` is set and `dotSelector` is not, then the plugin will
listen for clicks on `dotsContainer`'s children.
- If `dotsContainer` is not set, the plugin will create dots navigation
elements by itself (this has not changed) and will listen for clicks on the
created nav container's child elements